Can view a users recipes
Also added back in the logo
