+++ views/layouts/main.handlebars
@@ -61,7 +61,7 @@
           </li>
 
           <li class="nav-item">
-            <a class="nav-link" href="#">View Recipes</a>
+            <a class="nav-link" href="/recipes">View Recipes</a>
           </li>
 
           <li class="nav-item">
