+++ views/viewRecipes.handlebars
@@ -1,0 +1,16 @@
+  <main class="container-fluid mt-5">
+     <div class="row justify-content-around">
+      <div class="col-5 mt-1 mb-3">
+
+        <div class="row satBkgrnd justify-content-start rounded">
+          <div class="row satSmoke justify-content-center align-text-center cloud">
+            <p>Your Recipes</p>
+          </div>
+        </div>
+
+        {{#each recipesData as |recipe|}}
+        {{> recipecard}}
+        {{/each}}
+
+    </div>
+  </main>
